WIP: feat(vcs): new data model
* Updated the data model to accommodate the new generic approach to VCS
integration. This involves renaming the `github_...` tables to
`vcs_...`, adding a new column to the relevant tables to identify which
provider the records relate to, and more.

* Added an Alembic migration, including moving the repository data from
`oauthclient_remoteaccount` to the `vcs_repositories` table, which is a
complex and long-running operation. This will be supplemented by a
manual migration guide for instances like Zenodo where a several-minute
full DB lock is not feasible. The difference between whether to use the
automated migration or the manual one will be clarified in the docs.

* Added a repo-user m-to-m mapping table. By not storing repos in the
Remote Accounts table, we need a different way of associating users with
the repos they have access to. This table is synced using code that will
be included in other PRs.

* This PR contains only the data model changes themselves and not the
associated functional changes needed to do anything useful.

* This commit on its own is UNRELEASABLE. We will merge multiple commits
related to the VCS upgrade into the `vcs-staging` branch and then merge
them all into `master` once we have a fully release-ready prototype. At
that point, we will create a squash commit.