tests: compatibility for invenio-vcs
* Rewrote unit tests for compatibility with new invenio-vcs package by
making the tests provider-agnostic. This involved creating patches to
emulate the low-level behaviour of provider SDKs (for GitHub and GitLab)
based on an abstract interface. Future provider implementations will
also need corresponding test mock implementations.

* Generally renamed many references in the tests from invenio-github to
invenio-vcs. They should all currently be passing, but it's only
possible to fully test them if you have all the PRs for the new module
assembled. This is already done on the `master` branch of
https://github.com/palkerecsenyi/invenio-vcs for convenience.

* This commit on its own is UNRELEASABLE. We will merge multiple commits
related to the VCS upgrade into the `vcs-staging` branch and then merge
them all into `master` once we have a fully release-ready prototype. At
that point, we will create a squash commit.